Sortowanie
Wstępne posortowanie plików
