+++ projekt_byt/Login/logout.php
@@ -1,0 +1,7 @@
+<?php
+session_start();
+session_unset();
+session_destroy();
+header('Location: ../index.php');
+exit;
+?>
