+++ projekt_byt/Shop/manifest.json
@@ -1,0 +1,20 @@
+{
+  "name": "Formularz Dostawy i Podsumowanie Zakupów",
+  "short_name": "Zakupy",
+  "start_url": "/index.html",
+  "display": "standalone",
+  "background_color": "#f4f4f4",
+  "theme_color": "#28a745",
+  "icons": [
+    {
+      "src": "/icon-192x192.png",
+      "sizes": "192x192",
+      "type": "image/png"
+    },
+    {
+      "src": "/icon-512x512.png",
+      "sizes": "512x512",
+      "type": "image/png"
+    }
+  ]
+}
